+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Made mod less restrictive
 - Optimized imports
 - Minor code changes & formatting
+- Added GameTest & Unit tests